Merge comments into mono version.
The old modified Zarknorth copy of Neoforce was created from Mike Kukta's documented version. His website is no longer active, but I still have a modified version of his documentation. I wrote a quick tool to search for lines that are the same in the mono and documented versions. For each line that is the same, the tool checked for comments above it, and merged them into the source.